multi-pdf/requirements.txt
pdf2image==1.17.0
Pillow==10.2.0
boto3==1.34.69